Make admin-console deployable
