+++ builders/deploy.yaml
@@ -23,6 +23,27 @@ substitutions:
 
 steps:
 #
+# admin-console
+#
+- id: 'deploy-admin-console'
+  name: 'gcr.io/google.com/cloudsdktool/cloud-sdk:307.0.0-alpine'
+  args:
+  - 'bash'
+  - '-eEuo'
+  - 'pipefail'
+  - '-c'
+  - |-
+    gcloud run deploy "admin-console" \
+      --quiet \
+      --project "${PROJECT_ID}" \
+      --platform "managed" \
+      --region "${_REGION}" \
+      --image "gcr.io/${PROJECT_ID}/${_REPO}/admin-console:${_TAG}" \
+      --no-traffic
+  waitFor:
+  - '-'
+
+#
 # cleanup-exposure
 #
 - id: 'deploy-cleanup-exposure'
