+++ builders/promote.yaml
@@ -23,6 +23,26 @@ substitutions:
 
 steps:
 #
+# admin-console
+#
+- id: 'promote-admin-console'
+  name: 'gcr.io/google.com/cloudsdktool/cloud-sdk:307.0.0-alpine'
+  args:
+  - 'bash'
+  - '-eEuo'
+  - 'pipefail'
+  - '-c'
+  - |-
+    gcloud run services update-traffic "admin-console" \
+      --quiet \
+      --project "${PROJECT_ID}" \
+      --platform "managed" \
+      --region "${_REGION}" \
+      --to-revisions "${_REVISION}=${_PERCENTAGE}"
+  waitFor:
+  - '-'
+
+#
 # cleanup-exposure
 #
 - id: 'promote-cleanup-exposure'
