+++ cmd/admin-console/templates/authorizedapp.html
