+++ cmd/admin-console/templates/bottom.html
