+++ cmd/admin-console/templates/index.html
