+++ cmd/admin-console/templates/top.html
