+++ internal/admin/handle_signature_infos_test.go
@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC
+// Copyright 2021 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,34 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package authorizedapps
+package admin
 
 import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/google/exposure-notifications-server/internal/admin"
-	"github.com/google/exposure-notifications-server/internal/authorizedapp/model"
+	"github.com/google/exposure-notifications-server/internal/export/model"
 )
 
-func TestRenderAuthorizedApps(t *testing.T) {
-	// Hello developer!
-	// If this test fails, it's likely that you changed something in
-	//  internal/authorizedapp/model/
-	// And whatever you changed is used in the
-	//  tools/admin-console/templates/authorizedapp.html
-	// That is what caused the test failure.
-	m := admin.TemplateMap{}
-	authorizedApp := model.NewAuthorizedApp()
-	m["app"] = authorizedApp
+func TestRenderSignatureInfo(t *testing.T) {
+	m := TemplateMap{}
+	sigInfo := &model.SignatureInfo{}
+	m["siginfo"] = sigInfo
 
 	recorder := httptest.NewRecorder()
-	config := admin.Config{
-		TemplatePath: "../../../tools/admin-console/templates",
+	config := Config{
+		TemplatePath: "../../cmd/admin-console/templates",
 		TopFile:      "top",
 		BotFile:      "bottom",
 	}
-	err := config.RenderTemplate(recorder, "authorizedapp", m)
+	err := config.RenderTemplate(recorder, "siginfo", m)
 	if err != nil {
 		t.Fatalf("error rendering template: %v", err)
 	}
